requirements.txt
watchdog~=3.0.0
gitpython==3.1.32
schedule==1.2.0